+++ CHANGELOG.md
@@ -6,6 +6,8 @@
   * For `CopyMessage` order of parameters has been changed
 * `API`:
   * For `copyMessage` order of parameters has been changed
+* `Utils`:
+  * New extension `Message#sameMessage`
 
 ## 2.1.3